remove srun in sbatch script

+++ workflows/prepare_mRNA_dataset_part2.sh
@@ -54,14 +54,15 @@ for chr in {1..22} X
 do
   samtools view chr${chr}/sorted.chr${chr}.bam | cut -f1 > chr${chr}/read_ids.txt
   mkdir -p chr${chr}/fast5
-  srun -c 40 --mem 64GB -o ${HOME}/slurm/fast5_subset_chr${chr}.out -e ${HOME}/slurm/fast5_subset_chr${chr}.err \
-  fast5_subset -t 36 -i fast5 -s chr${chr}/fast5 -l chr${chr}/read_ids.txt &
+  fast5_subset -t 36 -i fast5 -s chr${chr}/fast5 -l chr${chr}/read_ids.txt
 done
 
 #########################################################################################################################
 #### clean up ###########################################################################################################
 #########################################################################################################################
-rm -rf ${WORKSPACE}/part*
-rm ${WORKSPACE}/fast5_paths_all
+if [ -f "${WORKSPACE}"/basecall_merged.fasta ]; then
+  rm -rf ${WORKSPACE}/part*
+  rm ${WORKSPACE}/fast5_paths_all
+fi
 
 echo "Finished on ${WORKSPACE}"
